[VectorOfTest]: fixed warning

+++ tests/libYARP_sig/VectorOfTest.cpp
@@ -245,7 +245,7 @@ public:
 
         //compare vector and tmp
         success = true;
-        if (tmp2.size() != vector.size())
+        if (tmp2.size() != (int)vector.size())
         {
             success = false;
         }
